Himalayan_Dairy_V2/requirements.txt
asgiref==3.11.0
beautifulsoup4==4.14.3
Django==4.2.27
django-bootstrap-v5==1.0.11
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
soupsieve==2.8
sqlparse==0.5.4
typing_extensions==4.15.0
whitenoise==6.11.0
